Finalize penalty card functionality
We now have buttons for tracking yellow/red cards issued to the
fencers. An icon shows up for an issued card, and clicking on that
icon removes it. As the goal is only to help the referee keep track of
what has happened in the bout so far, there is no need for a black card
button.

Passivity cards will be added later.